requirements.txt
flet
requests
paho-mqtt>=2.0.0
